+++ NEWS.md
@@ -5,6 +5,13 @@
 
 * Include directories in the file list
 
+* Always store relative file names in the archive
+
+* Strip leading `.` and `..` paths from the beginning of the
+  files in the archive
+
+* Fix reporting the size of large files in the archive
+
 # 1.0.0
 
 First public release.