+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# 1.0.0.9000
+# 2.0.0
 
 * New `zipr()` and `zipr_append()`, they always store relative file names
   in the archive.